Add tag button now works in edit mode. Styling updates.

+++ controllers/edit.js
@@ -33,7 +33,8 @@ router.post('/confirm', function(req, res) {
     outlets: parseInt(info.outlets),
     seating: parseInt(info.seating)
   };
-
+debugger;
+//Updates all radio button settings in edit window//
   db.placeinfo.find({where: {id: info.id}}).then(function(place) {
     db.placeinfo.update(
       {wifi: place.wifi + parseInt(info.wifi)},
@@ -56,41 +57,80 @@ router.post('/confirm', function(req, res) {
       {where: {id: info.id}}
     );
 
+//Creates all new tags from edit screen//
     if (typeof info.food == 'object') {
-    info.food.forEach(function(food) {
-      db.food.destroy({where: {food: food} } ).spread(function(foods, created) {
-        db.placeinfoFoods.destroy({where: {placeinfoId: info.id, foodId: foods.id} } );
+    fav.food.forEach(function(food) {
+      db.food.findOrCreate({where: {food: food} } ).spread(function(foods, created) {
+        db.placeinfoFoods.findOrCreate({where: {placeinfoId: user.id, foodId: foods.id} } );
       });
     });
   } else if (typeof info.food == 'string') {
-      db.food.destroy({where: {food: info.food} } ).spread(function(foods, created) {
-        db.placeinfoFoods.destroy({where: {placeinfoId: info.id, foodId: foods.id} } );
-      });
-    }
+    db.food.findOrCreate({where: {food: info.food} } ).spread(function(foods, created) {
+      db.placeinfoFoods.findOrCreate({where: {placeinfoId: user.id, foodId: foods.id} } );
+    });
+  }
 
     if (typeof info.drink == 'object') {
-    info.drink.forEach(function(drink) {
-      db.drink.destroy({where: {drink: drink} } ).spread(function(drinks, created) {
-        db.placeinfoDrinks.destroy({where: {placeinfoId: info.id, drinkId: drinks.id} } );
+    fav.drink.forEach(function(drink) {
+      db.drink.findOrCreate({where: {drink: drink} } ).spread(function(drinks, created) {
+        db.placeinfoDrinks.findOrCreate({where: {placeinfoId: user.id, drinkId: drinks.id} } );
       });
     });
   } else if (typeof info.drink == 'string') {
-      db.drink.destroy({where: {drink: info.drink} } ).spread(function(drinks, created) {
-        db.placeinfoDrinks.destroy({where: {placeinfoId: info.id, drinkId: drinks.id} } );
+      db.drink.findOrCreate({where: {drink: info.drink} } ).spread(function(drinks, created) {
+        db.placeinfoDrinks.findOrCreate({where: {placeinfoId: user.id, drinkId: drinks.id} } );
       });
     }
 
     if (typeof info.tag == 'object') {
-    info.tag.forEach(function(tag) {
+    fav.tag.forEach(function(tag) {
+      db.tag.findOrCreate({where: {tag: tag} } ).spread(function(tags, created) {
+        db.placeinfoTags.findOrCreate({where: {placeinfoId: user.id, tagId: tags.id} } );
+      });
+    });
+  } else if (typeof info.tag == 'string') {
+      db.tag.findOrCreate({where: {tag: info.tag} } ).spread(function(tags, created) {
+        db.placeinfoTags.findOrCreate({where: {placeinfoId: user.id, tagId: tags.id} } );
+      });
+    }
+
+//Destroys all selected tags in edit scree//
+    if (typeof info.foods == 'object') {
+    info.foods.forEach(function(food) {
+      db.food.destroy({where: {food: food} } ).spread(function(foods, created) {
+        db.placeinfoFoods.destroy({where: {placeinfoId: info.id, foodId: foods.id} } );
+      });
+    });
+  } else if (typeof info.foods == 'string') {
+      db.foods.destroy({where: {food: info.foods} } ).spread(function(foods, created) {
+        db.placeinfoFoods.destroy({where: {placeinfoId: info.id, foodId: foods.id} } );
+      });
+    }
+
+    if (typeof info.drinks == 'object') {
+    info.drinks.forEach(function(drink) {
+      db.drink.destroy({where: {drink: drink} } ).spread(function(drinks, created) {
+        db.placeinfoDrinks.destroy({where: {placeinfoId: info.id, drinkId: drinks.id} } );
+      });
+    });
+  } else if (typeof info.drinks == 'string') {
+      db.drink.destroy({where: {drink: info.drinks} } ).spread(function(drinks, created) {
+        db.placeinfoDrinks.destroy({where: {placeinfoId: info.id, drinkId: drinks.id} } );
+      });
+    }
+
+    if (typeof info.tags == 'object') {
+    info.tags.forEach(function(tag) {
       db.tag.destroy({where: {tag: tag} } ).spread(function(tags, created) {
         db.placeinfoTags.destroy({where: {placeinfoId: info.id, tagId: tags.id} } );
       });
     });
-  } else if (typeof info.tag == 'string') {
-      db.tag.destroy({where: {tag: info.tag} } ).spread(function(tags, created) {
+  } else if (typeof info.tags == 'string') {
+      db.tag.destroy({where: {tag: info.tags} } ).spread(function(tags, created) {
         db.placeinfoTags.destroy({where: {placeinfoId: info.id, tagId: tags.id} } );
       });
     }
+
     res.redirect('../favorites');
   });
 });
